[FIX] GCC 6 warning about placement-new operator
GCC 6 is able to detect when you use a placemente new of an object in a
space that is too small to contain it.

Unfortunately we get a false positive for:

    if (detail::any_::get_table<T>::is_small)
      new (&object) T(x);
    else
      object = new T(x);

where the warning is triggered for the false branch of the constant
expression.

So split any(const T &) constructor in two pairs and call the
appropriate function by tag dispatch.