[REF] Reverting vita::any to a std::any-like basic type
Previous implementation of vita::any (almost) supported a pseudo regular
type.

A *regular* type is a type that can be constructed and assigned-to and
assigned-from via copy or move, can be destroyed, and can be compared
equal-to. It can also be constructed from no argument (see
Alexander Stepanov's [Fundamentals of Generic Programming][1] and
[What is a regular type][2]).

`vita::any` also had support for a few other operations that are highly
useful in various `std` algorithms and containers:

- `std::hash`;
- `operator<<` and `operator>>`.

Both this features can be simulated with various techniques (e.g.
[cout map with boost::any][3], [boost::any to define my own print][4])

but, actually, we don't need a regular type.

Changing `basic_gene.par` from `terminal::param_t` to `vita::any` isn't
very useful: essentially the parameter is used to store a number, so the
`double` type already offers enough flexibility (it can store integer
numbers without loss of precision in a [large range][5]).

The "only" use left for `any` is for the `interpreter` class. There a
simple "copyable any" is all we need.

References issue Ticket #9

[1]: http://www.stepanovpapers.com/DeSt98.pdf
[2]: https://stackoverflow.com/q/13998945/3235496
[3]: https://stackoverflow.com/q/34673807/3235496
[4]: https://stackoverflow.com/q/4985351/3235496
[5]: https://stackoverflow.com/a/12445693/3235496