[REF] Compatibility changes to test/CMakeList.txt
It was using the `BOOST_WHATEVER_DYN_LINK` define and it's now using
`BOOST_ALL_NO_LIB`.

See https://stackoverflow.com/q/28887680

References:

BOOST_WHATEVER_DYN_LINK
Forces library "whatever" to be linked as a dll rather than a static library
on Microsoft Windows: replace the WHATEVER part of the macro name with the
name of the library that you want to dynamically link to, for example use
BOOST_DATE_TIME_DYN_LINK or BOOST_REGEX_DYN_LINK etc (this macro is used to
turn on __declspec(dllimport) modifiers, so that the compiler knows which
symbols to look for in a dll rather than in a static library). Note that
there may be some libraries that can only be statically linked (Boost.Test
for example) and others which may only be dynamically linked (Boost.Thread
for example), in these cases this macro is unsupported.

BOOST_ALL_NO_LIB
Forces all libraries that have separate source, to be linked as dll's
rather than static libraries on Microsoft Windows (this macro is used to
turn on __declspec(dllimport) modifiers, so that the compiler knows which
symbols to look for in a dll rather than in a static library). Note that
there may be some libraries that can only be statically linked (Boost.Test
for example) and others which may only be dynamically linked (Boost.Thread
for example), in these cases this macro has no effect.