[FIX] GCC 7 warning about placement-new operator
GCC 7 is able to detect when you use a placement new of an object in a
space that is too small to contain it.

Unfortunately we get a false positive for the any::assign(const T &)
function.

So split assign(const T &) in two pairs and call the appropriate function
by tag dispatch.

This is the same escamotage used for GCC6 and the any constructor...
perhaps GCC is becoming too smart?