[REF] P(terminal) = P(function) for symbol_set::roulette()
The probability of extraction is the same for terminals functions
(regardeless of the weights of the symbols).

If all symbols have the same probability of appearing into a chromosome,
there could be some problems.
For instance, if our problem has many variables (let's say 100) and the
function set has only 4 symbols we cannot get too complex trees because the
functions have a reduced chance to appear in the chromosome (e.g. it
happens in the Forex example).