[REF] Removing src_problem(const char []) delegating constructor
The delegating constructor was needed to avoid that the
`src_problem("data.csv")` call call could be resolved with
`src_problem(bool)`.

Conversion from `const char []` to `bool` is a standard conversion, while
the one to `std::string` is a user-defined conversion.

Since the src_problem(bool) constructor was removed a long time ago
we don't need the delegating constructor anymore.