[MAK] Now using `CONFIGURE_DEPENDS` with CMake
With code like:

```
file(GLOB sources "*.cc")
```

adding a new .cc file to the directory will not trigger a re-configure in an incremental build. This is bad behavior because it forces a developer to re-run CMake as opposed to just the build tool.

One solution is to use `CONFIGURE_DEPENDS`, which will cause the generated build to re-evaluate the globs and re-configure if anything changes. This code correctly sets dependencies.

```
file(GLOB sources CONFIGURE_DEPENDS "*.cc")
```