[REF] Remove `thread_local` specifier from random engine
Don't pay for what you don't use is a philosophy of the design of the C++.

Vita doesn't support concurrency and the project is in maintenance mode (only minor modifications / bug fixing; no new features).

Concurrency won't be implemented so we can remove unused performance reducing features.

The new ULTRA framework (https://github.com/morinim/ultra) is the heir-project of VITA and it supports concurrency from the beginning.